Move the HTTP test client to `test_common`
`test_response` is a plain JSON over HTTP client with no knowledge of
the RPC API, but it lived in `rpc_test` and pulled in the IPC server
and request processor headers it never used. It moves to `test_common`
with only the includes it needs, so the upcoming `rpc_server` unit
tests in the core suite can drive the server with the same client the
RPC API tests use.
